requirements/production.txt
-r common.txt
boto==2.6.0
cssmin==0.1.4
django-heroku-memcacheify==0.3
django-heroku-postgresify==0.2
django-storages==1.1.4
gevent==0.13.8
gunicorn==0.16.1
jsmin==2.0.2
newrelic==1.8.0.13
psycopg2==2.4.5
pylibmc==1.2.3
